crimeSuccessChance: Switch to formulas
With BitBurner v2.2.0 ns.sleeve.getInformation() was removed and thus
the calculation of the crimeSuccessChance for sleeves would have failed.
But a new formula was added to calculate the crimeSuccessChance for the
player or a sleeve. So the calculation was refactored to use the formula
and the manual calculation as well as calculation data basis could be
removed.