hacknet: Rework spending hashes for corp funds
The getCorporation() function is quite RAM heavy and can't be used at
the beginning of a BitNode. So we can't stop spending hashes for
corporation funds as soon as the corporation went public. Therefore this
check is removed but we only spend hashes for corporation funds 10 times
per run instead of 100 times.